refactor(server): split match.ts — extract 3 readonly handlers (S27.8.17)
Second slice on `apps/server/src/routes/match.ts`. Extract the 3
remaining read-only handlers (`handleListMatchTurns`,
`handleGetMatchResults`, `handleListLiveMatches`) into a dedicated
module `routes/match-readonly-handlers.ts` (302 lines).

These handlers cover:
- `GET /:id/turns` — turn history with snapshots (number, type, half,
  turn, score, moveType). Player-only check.
- `GET /:id/results` — final score, winner, ELO, team stats, fan
  attendance for ended matches.
- `GET /live` — 20 most recent active matches for spectator mode.

All thematically read-only and isolated from the lifecycle handlers
that remain in `match.ts`.

Re-export from `match.ts` to preserve the public API consumed by
integration tests.

Leaf dependencies only (`prisma`, `getSpectatorCount`, `serverLog`,
`sendError/sendSuccess`, `AuthenticatedRequest`). No cycle.

`match.ts` shrinks 1631 -> 1393 lines (-238 net lines, cumulative
-373 vs origin 1766 since S27.8.16).

- 1787/1787 server tests pass (no regression)
- typecheck clean

Next slice will tackle `handleGetMatchState` (~685 lines), the
remaining heavyweight in `match.ts`.
